soc_estimation/lstm/requirements.txt
flask>=2.0.0
tensorflow-cpu>=2.15.0
numpy>=1.21.0
